Modify score to float, and add GET

+++ src/api/model/training.py
@@ -131,9 +131,9 @@ class BlockStatus(BaseModel):
 
 
 class ScoreBody(BaseModel):
-    score: int = Field(
+    score: float = Field(
         title="Score",
         description="Score given to training by user",
-        ge=0,
-        le=5,
+        ge=0.0,
+        le=5.0,
     )
